Fixed use of doubles instead of int64_t.
Former-commit-id: 38a753037154e35981d687a4f6671a5d641caa90 [formerly 3c8d55c549eaa951be8a5ad8aa3aa3ce4d2e195c] [formerly 15000d5eb8035bda06d944aa04c7fedf94d97e17] [formerly b81797926558c71f70295bd89461932b079e8a73] [formerly 5f01f84d222742c11df565ccbe302310a3b88307 [formerly 3689d9f0942f4dbacabeffd92039b3dfd57e2b6a [formerly 6908411f4a37adc9e126e04fa50bfa07c661feac] [formerly 5f01f84d222742c11df565ccbe302310a3b88307 [formerly 5f01f84d222742c11df565ccbe302310a3b88307 [formerly 5f01f84d222742c11df565ccbe302310a3b88307 [formerly d9dffba62566bbb908b6edcc0ad842f695a444d7]]]]]] [formerly 5b65de1c684a8a75a9b52d5391fe90083cbe5ef3 [formerly 71366d464a12aa6b9209f215dabf1afc15f1c864] [formerly d9385e399735004d91a518a883b961a35385790a] [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51] [formerly 6e746b905fbddc962aca96f558d2ccda0a950e8a [formerly bb5609202e9919cf2fc2daa106c37bdfb3a67060] [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly eb55a6f37c643bdf0fdb4db32aaa6c988d3114c3]]]]] [formerly fb7fb59b3f14e57b42ef429f70d5740e7d2332db [formerly fb7fb59b3f14e57b42ef429f70d5740e7d2332db [formerly fb7fb59b3f14e57b42ef429f70d5740e7d2332db [formerly fb7fb59b3f14e57b42ef429f70d5740e7d2332db [formerly 3a007b1aebac035c75123518669e66614ff2fcb7]]]]]]
Former-commit-id: 5b65de1c684a8a75a9b52d5391fe90083cbe5ef3 [formerly 71366d464a12aa6b9209f215dabf1afc15f1c864] [formerly d9385e399735004d91a518a883b961a35385790a] [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51] [formerly 6e746b905fbddc962aca96f558d2ccda0a950e8a [formerly bb5609202e9919cf2fc2daa106c37bdfb3a67060] [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly 97b4c0c4bf673a1eeae7a583695a0d10b1887a51 [formerly eb55a6f37c643bdf0fdb4db32aaa6c988d3114c3]]]]]
Former-commit-id: ee7a548f9116ec48f3462bbf4bca3fc8f36352d1 [formerly 397ee2de8af3297f012386d92225791feb633902 [formerly d9c437b0954e5187707f1a7c6b3bdb608cf5d7b9] [formerly ee7a548f9116ec48f3462bbf4bca3fc8f36352d1 [formerly ee7a548f9116ec48f3462bbf4bca3fc8f36352d1 [formerly ee7a548f9116ec48f3462bbf4bca3fc8f36352d1 [formerly 95ac6a572fa36aa358b97c4834a1da2802b8a736]]]]]
Former-commit-id: 8e5c715c60c30744c9f4df16db08728bdcf5beb6
Former-commit-id: 6913652a17b045a27ca17589b5a9df780b4e6aaf
Former-commit-id: bb2ed19268f2d3b08fdb972206fd33c9cab79493
Former-commit-id: 7388917941ac4a216572b07966003c134ef477c7
Former-commit-id: d72f4594532c4bff448408230512172b93917ed2

+++ mrViewer/src/gui/mrvReel.cpp
@@ -116,13 +116,13 @@ mrv::media Reel_t::media_at( const boost::int64_t f ) const
     mrv::MediaList::const_iterator e = images.end();
 
     mrv::media m = images.front();
-    double mn = m->position();
+    int64_t mn = m->position();
 
     m = images.back();
-    double mx = m->position() + m->image()->duration() - 1;
+    int64_t mx = m->position() + m->image()->duration() - 1;
     if ( mn > mx ) 
     {
-       double t = mx;
+       int64_t t = mx;
        mx = mn; mn = t;
     }
